backend/pyproject.toml
[tool.poetry]
name = "MaterialManager"
version = "0.1.0"
description = "A manager for materials"
authors = ["xiaoyezi <1372755472@foxmail.com>"]
license = "MIT"
readme = "../README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.100.0"
uvicorn = "^0.23.1"
tortoise-orm = {extras = ["asyncmy"], version = "^0.20.0"}
pydantic = "^2.3.0"
email-validator = "^2.0.0.post2"
passlib = "^1.7.4"
pyjwt = "^2.7.0"
black = "^23.7.0"
isort = "^5.12.0"
ruff = "^0.0.281"
loguru = "^0.7.0"
pydantic-settings = "^2.0.3"
argon2-cffi = "^23.1.0"
python-dotenv = "^1.0.1"
cryptography = "^42.0.5"
ruamel-yaml = "^0.18.6"
pymysql = "^1.1.0"
python-multipart = "^0.0.9"

[tool.black]
line-length = 120
target-version = ["py310", "py311"]

[[tool.poetry.source]]
name = "tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple/"
priority = "primary"


[[tool.poetry.source]]
name = "PyPI"
priority = "primary"

[tool.ruff]
line-length = 120
extend-select = [
  # "I",    # isort
#   "B",    # flake8-bugbear
#   "C4",   # flake8-comprehensions
#   "PGH",  # pygrep-hooks
  # "RUF",  # ruff
#   "W",    # pycodestyle
#   "YTT",  # flake8-2020
]
ignore = [
    "F403",
    "F405",
]